etf-intelligence/backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
python-dotenv
yfinance
apscheduler
pytz
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
python-multipart
email-validator
httpx
pytest
pytest-asyncio
pydantic[email]